requirements.txt
tensorflow>=2.10.0
streamlit>=1.28.0
numpy
matplotlib
seaborn
scikit-learn
plotly
pillow>=9.0.0
